+++ source/patterns/@aws-solutions-constructs/aws-alb-lambda/test/integ.privateApiNewResources.ts
@@ -33,7 +33,7 @@ stack.templateOptions.description = 'Integration Test for HTTP API with a new fu
 const props: AlbToLambdaProps = {
   lambdaFunctionProps: {
     code: lambda.Code.fromAsset(`${__dirname}/lambda`),
-    runtime: lambda.Runtime.NODEJS_12_X,
+    runtime: lambda.Runtime.NODEJS_14_X,
     handler: 'index.handler'
   },
   listenerProps: {