+++ source/patterns/@aws-solutions-constructs/aws-eventbridge-kinesisfirehose-s3/test/integ.eventbridge-kinesisfirehose-s3-existing-eventbus.ts
@@ -21,7 +21,7 @@ import * as s3 from "@aws-cdk/aws-s3";
 const app = new App();
 const stack = new Stack(app, generateIntegStackName(__filename));
 stack.templateOptions.description = 'Integration Test for aws-eventbridge-kinesisfirehose-s3';
-const existingEventBus = new events.EventBus(stack, `test-existing-eventbus`, {});
+const existingEventBus = new events.EventBus(stack, `test-existing-eventbus`, { eventBusName: 'test' });
 const props: EventbridgeToKinesisFirehoseToS3Props = {
   eventRuleProps: {
     eventPattern: {