+++ source/patterns/@aws-solutions-constructs/aws-events-rule-kinesisfirehose-s3/test/integ.events-rule-kinesisfirehose-s3-new-eventbus.ts
@@ -27,7 +27,7 @@ const props: EventsRuleToKinesisFirehoseToS3Props = {
       source: ['solutionsconstructs']
     }
   },
-  eventBusProps: {},
+  eventBusProps: { eventBusName: 'test' },
   bucketProps: {
     removalPolicy: RemovalPolicy.DESTROY
   },