+++ source/patterns/@aws-solutions-constructs/aws-events-rule-lambda/test/integ.eventsrule-existing-eventbus.ts
@@ -21,12 +21,12 @@ import { generateIntegStackName } from '@aws-solutions-constructs/core';
 const app = new App();
 const stack = new Stack(app, generateIntegStackName(__filename));
 
-const existingEventBus = new events.EventBus(stack, `existing-event-bus`);
+const existingEventBus = new events.EventBus(stack, `existing-event-bus`, {  eventBusName: 'test'  });
 
 const props: EventsRuleToLambdaProps = {
   lambdaFunctionProps: {
     code: lambda.Code.fromAsset(`${__dirname}/lambda`),
-    runtime: lambda.Runtime.NODEJS_12_X,
+    runtime: lambda.Runtime.NODEJS_14_X,
     handler: 'index.handler'
   },
   existingEventBusInterface: existingEventBus,