+++ source/patterns/@aws-solutions-constructs/aws-lambda-ssmstringparameter/test/integ.deployFunction.ts
@@ -25,7 +25,7 @@ stack.templateOptions.description = 'Integration Test for aws-lambda-ssmstringpa
 // Definitions
 const props: LambdaToSsmstringparameterProps = {
   lambdaFunctionProps: {
-    runtime: lambda.Runtime.NODEJS_10_X,
+    runtime: lambda.Runtime.NODEJS_14_X,
     handler: 'index.handler',
     code: lambda.Code.fromAsset(`${__dirname}/lambda`)
   },