+++ source/patterns/@aws-solutions-constructs/aws-lambda-stepfunctions/test/integ.existing-function.ts
@@ -28,7 +28,7 @@ const startState = new stepfunctions.Pass(stack, 'StartState');
 
 // Setup the "existing" Lambda function props
 const lambdaFunctionProps = {
-  runtime: lambda.Runtime.NODEJS_10_X,
+  runtime: lambda.Runtime.NODEJS_14_X,
   handler: 'index.handler',
   code: lambda.Code.fromAsset(`${__dirname}/lambda`)
 };